# Conflicts:
#	Assets/Scenes/Main Menu.unity
#	Assets/Scripts/MovePlayer.cs
#	UserSettings/Layouts/default-2021.dwlt

+++ Assets/Scripts/SaveSystem.cs
@@ -1,11 +1,12 @@
-
+using System.Collections.Generic;
+using System.Collections;
 using System.IO;
 using System.Runtime.Serialization.Formatters.Binary;
 using UnityEngine;
 
 public static class SaveSystem
 {
-    public static void SaveGame(float distance)
+    public static void SaveGame(int distance)
     {
         BinaryFormatter formatter = new BinaryFormatter();
         string path = Application.persistentDataPath + "/Save.txt";
@@ -14,8 +15,5 @@ public static class SaveSystem
         formatter.Serialize(stream,distance);
         stream.Close();
     }
-    
-
-
 
 }
